pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "travelist-backend"
version = "0.0.1"
description = "Travelist+ FastAPI backend"
readme = "README.md"
requires-python = ">=3.12"
authors = [
  {name = "Travelist Team"}
]
dependencies = [
  "fastapi>=0.111.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic-settings>=2.3.0",
  "langgraph>=0.2.21",
  "langchain-core>=0.3.0",
  "psycopg[binary]>=3.2.0",
  "redis>=5.0.0",
  "sqlalchemy>=2.0.0",
  "jinja2>=3.1.0",
  "geoalchemy2>=0.14.2",
  "alembic>=1.13.0",
  "httpx>=0.27.0",
  "requests>=2.32.0",
]

[project.optional-dependencies]
dev = [
  "black>=24.4.0",
  "ruff>=0.5.0",
  "pytest>=8.2.0",
  "pytest-cov>=5.0.0",
  "httpx>=0.27.0",
  "alembic>=1.13.0",
]

[tool.setuptools.packages.find]
where = ["backend"]

[tool.black]
line-length = 88
target-version = ["py312"]

[tool.ruff]
line-length = 88
src = ["backend/app", "backend/tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "B"]

[tool.pytest.ini_options]
testpaths = ["backend/tests"]
addopts = "--maxfail=1 --disable-warnings"
asyncio_default_fixture_loop_scope = "function"

[tool.coverage.run]
source = ["backend/app"]
omit = ["*/__init__.py"]
branch = true

[tool.coverage.report]
show_missing = true
skip_covered = true
